+++ talpid-core/src/routing/mod.rs
@@ -17,6 +17,9 @@ use netlink_packet_route::rtnl::constants::RT_TABLE_MAIN;
 
 pub use imp::{Error, RouteManager};
 
+#[cfg(target_os = "linux")]
+pub use imp::RouteManagerCommand;
+
 /// A netowrk route with a specific network node, destinaiton and an optional metric.
 #[derive(Debug, Hash, Eq, PartialEq, Clone)]
 pub struct Route {