+++ talpid-core/src/tunnel/openvpn.rs
@@ -6,7 +6,10 @@ use crate::{
         stoppable_process::StoppableProcess,
     },
     proxy::{self, ProxyMonitor, ProxyResourceData},
+    routing,
 };
+#[cfg(target_os = "linux")]
+use futures::channel::oneshot;
 use std::{
     collections::HashMap,
     fs,
@@ -36,6 +39,10 @@ pub enum Error {
     /// Failed to initialize the tokio runtime.
     #[error(display = "Failed to initialize the tokio runtime")]
     RuntimeError(#[error(source)] io::Error),
+
+    /// Failed to set up routing.
+    #[error(display = "Failed to setup routing")]
+    SetupRoutingError(#[error(source)] routing::Error),
 
     /// Unable to start, wait for or kill the OpenVPN process.
     #[error(display = "Error in OpenVPN process management: {}", _0)]
@@ -145,6 +152,7 @@ impl OpenVpnMonitor<OpenVpnCommand> {
         params: &openvpn::TunnelParameters,
         log_path: Option<PathBuf>,
         resource_dir: &Path,
+        #[cfg(target_os = "linux")] route_manager: &mut routing::RouteManager,
     ) -> Result<Self>
     where
         L: Fn(TunnelEvent) + Send + Sync + 'static,
@@ -163,7 +171,22 @@ impl OpenVpnMonitor<OpenVpnCommand> {
             _ => None,
         };
 
-        let on_openvpn_event = move |event, env| {
+        #[cfg(target_os = "linux")]
+        let route_manager_tx = route_manager.channel().map_err(Error::SetupRoutingError)?;
+
+        let on_openvpn_event = move |event, env: HashMap<String, String>| {
+            #[cfg(target_os = "linux")]
+            if event == openvpn_plugin::EventType::Up {
+                let (tx, rx) = oneshot::channel();
+                let interface = env.get("dev").unwrap().to_owned();
+                route_manager_tx
+                    .unbounded_send(routing::RouteManagerCommand::SetTunnelLink(interface, tx))
+                    .unwrap();
+                tokio::task::block_in_place(move || {
+                    futures::executor::block_on(rx).unwrap();
+                });
+                return;
+            }
             if event == openvpn_plugin::EventType::RouteUp {
                 // The user-pass file has been read. Try to delete it early.
                 let _ = fs::remove_file(&user_pass_file_path);